use the DependentSecondPass abstraction to sort FkSecondPasses
Or some of them, at least. The FkSecondPasses which refer to a non-primary-key field (endOfQueueFkSecondPasses) don't seem to lend themselves nicely to this, since 'dependentUpon()` would require information that isn't available until that second pass runs. That's probably why this is currently implemented with RecoverableException, and I don't really know a better way to do this.

The primary-key FkSecondPasses, though, can pretty easily implement DependentSecondPass, so we can use this instead of `buildRecursiveOrderedFkSecondPasses()`